JAVA:*JAVA-OBJECT-TO-STRING-LENGTH* controls pretty print length.
The default of 32 characters is the length beyond which the toString()
output from which an otherwise unspecialized JAVA-OBJECT elides its
representation.

Changed the character sequence indicating elision from "..." to "...."
to avoid confusing the user that any of the pretty printer variables
are responsible for an elision.

Setting this variable to NIL disables truncation of any kind.

git-svn-id: http://abcl.org/svn/trunk/abcl@13277 1c010e3e-69d0-11dd-93a8-456734b0d56f