Upgrade to asdf-3.1.6
N.b. one will have to do a full clean of existing fasls for this
change to compile.  Otherwise one gets errors about "wrong number of
arguments for DEFSETF".

N.b. the configuration directory under Windows has changed in
incompatible ways.  We intend to "fix" this in a backwards compatible
manner before releasing a version of ABCL based on this change but
wish to get asdf-3.1.6 out into circulation as a first step.

git-svn-id: http://abcl.org/svn/trunk/abcl@14838 1c010e3e-69d0-11dd-93a8-456734b0d56f